Migrate dumb presentational components from web app into @repo/ui
Moves 11 purely presentational components (Button, ActionBar, ActionSeparator,
ActionText, ActionIcon, Method, Status, TrueFalseMark, ToggleSwitch, DateTime,
Url) out of apps/web into the shared @repo/ui package so they can be reused
across apps and packages.

- DateTime and Url are refactored to accept pre-processed props (no helper deps)
- Turborepo starter placeholder button renamed to turbo-button
- All web app call sites updated to import from @repo/ui/*
