--- apps/web/app/components/DateTime.tsx
@@ -1,15 +1,0 @@
-import { formatDateTime } from "../helpers/helpers";
-
-export function DateTime({
-  dateTime,
-  timeOnly = false,
-}: {
-  dateTime: string;
-  timeOnly: boolean;
-}) {
-  return (
-    dateTime && (
-      <div className="ml-auto">{formatDateTime(dateTime, timeOnly)}</div>
-    )
-  );
-}
