+++ apps/web/app/components/KeepPagesToggle.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ToggleSwitch from "./ToggleSwitch";
+import ToggleSwitch from "@repo/ui/toggle-switch";
 
 export default function KeepPagesToggle(props: {
   isActive: boolean;
