+++ packages/ui/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "exports": {
     "./button": "./src/button.tsx",
+    "./turbo-button": "./src/turbo-button.tsx",
     "./card": "./src/card.tsx",
     "./code": "./src/code.tsx",
     "./text-content": "./src/text-content.tsx",
@@ -13,7 +14,17 @@
     "./magnifying-glass-svg": "./src/magnifying-glass-svg.tsx",
     "./line-clamp": "./src/line-clamp.tsx",
     "./expand-button": "./src/expand-button.tsx",
-    "./use-collapsed": "./src/use-collapsed.ts"
+    "./use-collapsed": "./src/use-collapsed.ts",
+    "./action-bar": "./src/action-bar.tsx",
+    "./action-separator": "./src/action-separator.tsx",
+    "./action-text": "./src/action-text.tsx",
+    "./action-icon": "./src/action-icon.tsx",
+    "./true-false-mark": "./src/true-false-mark.tsx",
+    "./method": "./src/method.tsx",
+    "./status": "./src/status.tsx",
+    "./toggle-switch": "./src/toggle-switch.tsx",
+    "./date-time": "./src/date-time.tsx",
+    "./url": "./src/url.tsx"
   },
   "scripts": {
     "lint": "eslint . --max-warnings 0",
