+++ packages/ui/src/date-time.tsx
@@ -1,0 +1,3 @@
+export function DateTime({ value }: { value: string }) {
+  return value ? <div className="ml-auto">{value}</div> : null;
+}
