+++ packages/ui/src/method.tsx
@@ -1,4 +1,5 @@
 import type React from "react";
+
 export const methodColors: { [key: string]: string } = {
   GET: "text-green-600",
   POST: "text-blue-500",
